Rename shape_calcurator's file to follow pep8 conventions
