+++ skl2onnx/shape_calculators/__init__.py
@@ -6,52 +6,52 @@
 
 # To register shape calculators for scikit-learn operators,
 # import associated modules here.
-from . import ArrayFeatureExtractor
-from . import Concat
-from . import DictVectorizer
-from . import Flatten
-from . import FunctionTransformer
-from . import Imputer
-from . import KBinsDiscretiser
-from . import KMeans
-from . import LabelBinariser
-from . import LabelEncoder
-from . import LinearClassifier
-from . import LinearRegressor
-from . import NearestNeighbours
-from . import OneHotEncoder
-from . import OneVsRestClassifier
-from . import PolynomialFeatures
-from . import Scaler
-from . import SVD
-from . import SVM
-from . import TextVectorizer
-from . import TfidfTransformer
-from . import VotingClassifier
-from . import ZipMap
+from . import array_feature_extractor
+from . import concat
+from . import dict_vectorizer
+from . import flatten
+from . import function_transformer
+from . import imputer
+from . import k_bins_discretiser
+from . import k_means
+from . import label_binariser
+from . import label_encoder
+from . import linear_classifier
+from . import linear_regressor
+from . import nearest_neighbours
+from . import one_hot_encoder
+from . import one_vs_rest_classifier
+from . import polynomial_features
+from . import scaler
+from . import svd
+from . import svm
+from . import text_vectorizer
+from . import tfidf_transformer
+from . import voting_classifier
+from . import zip_map
 
 __all__ = [
-    ArrayFeatureExtractor,
-    Concat,
-    DictVectorizer,
-    Flatten,
-    FunctionTransformer,
-    Imputer,
-    KBinsDiscretiser,
-    KMeans,
-    LabelBinariser,
-    LabelEncoder,
-    LinearClassifier,
-    LinearRegressor,
-    NearestNeighbours,
-    OneHotEncoder,
-    OneVsRestClassifier,
-    PolynomialFeatures,
-    Scaler,
-    SVD,
-    SVM,
-    TextVectorizer,
-    TfidfTransformer,
-    VotingClassifier,
-    ZipMap,
+    array_feature_extractor,
+    concat,
+    dict_vectorizer,
+    flatten,
+    function_transformer,
+    imputer,
+    k_bins_discretiser,
+    k_means,
+    label_binariser,
+    label_encoder,
+    linear_classifier,
+    linear_regressor,
+    nearest_neighbours,
+    one_hot_encoder,
+    one_vs_rest_classifier,
+    polynomial_features,
+    scaler,
+    svd,
+    svm,
+    text_vectorizer,
+    tfidf_transformer,
+    voting_classifier,
+    zip_map,
 ]
