+++ skl2onnx/shape_calculators/array_feature_extractor.py
