+++ skl2onnx/shape_calculators/dict_vectorizer.py
