+++ skl2onnx/shape_calculators/flatten.py
