+++ skl2onnx/shape_calculators/function_transformer.py
