+++ skl2onnx/shape_calculators/k_means.py
