+++ skl2onnx/shape_calculators/label_binariser.py
