+++ skl2onnx/shape_calculators/label_encoder.py
