+++ skl2onnx/shape_calculators/linear_classifier.py
