+++ skl2onnx/shape_calculators/linear_regressor.py
