+++ skl2onnx/shape_calculators/nearest_neighbours.py
