+++ skl2onnx/shape_calculators/one_hot_encoder.py
