+++ skl2onnx/shape_calculators/one_vs_rest_classifier.py
