+++ skl2onnx/shape_calculators/polynomial_features.py
