+++ skl2onnx/shape_calculators/scaler.py
