+++ skl2onnx/shape_calculators/svm.py
