+++ skl2onnx/shape_calculators/text_vectorizer.py
