+++ skl2onnx/shape_calculators/voting_classifier.py
