+++ skl2onnx/shape_calculators/zip_map.py
