+++ tests/test_sklearn_svm_converters.py
@@ -10,7 +10,7 @@ from sklearn import __version__ as sk__version__
 from skl2onnx import convert_sklearn, update_registered_converter
 from skl2onnx.common.data_types import FloatTensorType
 from skl2onnx.operator_converters.SVM import convert_sklearn_svm
-from skl2onnx.shape_calculators.SVM import calculate_sklearn_svm_output_shapes
+from skl2onnx.shape_calculators.svm import calculate_sklearn_svm_output_shapes
 from test_utils import dump_data_and_model
